+++ user-database/src/test/java/gradution_project/login_system/user_database/unit_test/APIExceptionHandlerUnitTest.java
@@ -1,0 +1,4 @@
+package gradution_project.login_system.user_database.unit_test;
+
+public class APIExceptionHandlerUnitTest {
+}
